requirements.txt
fastapi
pywa
sqlalchemy
asyncpg
python-dotenv
psycopg2-binary
httpx
uvicorn
pydantic
databases
alembic
pytest
pytest-asyncio
pytest-cov
requests
boto3
botocore